clean up, normalize scaling, fixed Set_Power_M?
